requirements.txt
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy=2.5.0
Faker=13.3.0
mixer=7.2.2
